Separation of json combiner (#47)
Separation of join_json_files, addressing #43
